Move balancer from generators to controls; add stub starting-point for AudioIter (name WIP) to represent a inite, predetermined track sample (e.g. a drum kick)
