+++ README.md
@@ -16,3 +16,9 @@ TODOs:
     - [ ] Envelope filters?
     - [ ] Band-pass filters?
 - [ ] Research and implement more controls
+- [ ] Implement metronome
+- [ ] Explore possibility of integration of 3rd-party effects (e.g. VST instruments)
+- ~~[ ] Sort out the `'static` situation (shouldn't be a requirement for `Generator`, `Filter` types)~~
+    - Currently prioritizing ease of use and functionality over correctness -- `'static` as a
+      requirement for `Generator` has not proven to be a roadblock in any way, and it may be
+      preferable to code littered with `<'a>` explicit lifetimes
